Update README.md
The objectives of this assignment are for you to practice: (1) the use of C++ I/O streams, includ-ing error handling, (2) writing a simple class with constructors, accessors and mutators, and (3) dynamic allocation and de-allocation of one-dimensional arrays. You will do so through the design of a program that parses drawing commands from the standard input, displaying appropriate error messages if necessary, and creating and maintaining objects that represent drawn shapes.

Problem Statement

The assignment consists of two main parts. In the first part, you will write a command parser that provides a textual input interface to your program. The parser takes a sequence of commands as input. The commands create, delete, modify and display shapes to be drawn on the screen. Each command consists of an operation keyword followed by arguments. The command and the arguments are separated by one or more spaces. Thus, the code you will write for this part of the assignment should take input from the standard input, parse it, verify that it is correct, print a response or error message and either create, delete or modify Shape objects that represent the shapes specified in the command. The command parser loops, processing input as long as input is available.

In the second part of the assignment, you will implement a simple “database” of objects that stores the created shapes. To do so, you will implement a class called Shape that is used to store the shape properties. Further, you will create and maintain a dynamically allocated array of pointers to Shape objects to keep track of Shape objects created and deleted.

In this assignment, you will not actually draw the shapes to the screen, just process the com-mands and maintain the database of the Shape objects.

String Streams